Dependencies to LASFields.h outside of the qPDALIO plugin :(
