+++ plugins/core/IO/qPDALIO/src/CMakeLists.txt
@@ -12,7 +12,6 @@ set( CC_PLUGIN_CUSTOM_HEADER_LIST
 set( CC_PLUGIN_CUSTOM_SOURCE_LIST
 	${CC_PLUGIN_CUSTOM_SOURCE_LIST} 
 
-	${CMAKE_CURRENT_SOURCE_DIR}/LASFields.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/LASFilter.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/LASOpenDlg.cpp
